+++ ndk/src/lib.rs
@@ -16,15 +16,15 @@
 #![warn(missing_debug_implementations, trivial_casts)]
 
 pub mod asset;
+pub mod audio;
 pub mod bitmap;
 pub mod configuration;
 pub mod event;
+pub mod hardware_buffer;
+pub mod hardware_buffer_format;
 pub mod input_queue;
 pub mod looper;
+pub mod media;
 pub mod native_activity;
 pub mod native_window;
-
-pub mod audio;
-pub mod hardware_buffer;
-pub mod media;
 pub mod trace;